Fix: notiSetting UserDefaults app Group으로 변경
